--- src/components/Badge/style.css
@@ -1,25 +1,0 @@
-.skeleton {
-  position: relative;
-  overflow: hidden;
-  background: #ccc;
-  color: rgb(244, 244, 244);
-}
-.skeleton::after {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: linear-gradient(90deg, #ccc, #dedede, #ccc);
-  animation: progress 1s ease-in-out infinite;
-}
-
-@keyframes progress {
-  0% {
-    transform: translate3d(-100%, 0, 0);
-  }
-  100% {
-    transform: translate3d(100%, 0, 0);
-  }
-}
