+++ stories/badge.js
@@ -1,20 +1,11 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { action } from '@storybook/addon-actions'
 
 import Badge from './../src/components/Badge'
 
 storiesOf('Badge', module)
-  .add('test', () => (
+  .add('standard', () => (
     <Badge
       children='GO !'
-      onClick={action('clicked !')}
-    />
-  ))
-  .add('[skeleton]', () => (
-    <Badge
-      children='GO !'
-      onClick={action('clicked !')}
-      isLoading
     />
   ))
